Provide infix:<eqv> comparison multis instead of infix:<eq>
In Perl 6 eq is string equality, while eqv is structural value equivalence,
so objects (and arrays of them) should be compared with eqv.  This change
also allows cleaning up some hackery needed by the original comparison
code.  Note that this commit does not include matching updates to the
tests, which are in separate commits.