Be consistent about explicit ! and ? for params in new() declarations
Fix the one place in which explicit ! and ? markers were not used on all
parameters in new() declarations.  These are not actually necessary when
the default ! is used for positional parameters or ? is used for named
parameters, but appear to be used intentionally as constructor usage
documentation throughout the code base, except for this one missing ?,
so fix the inconsistency.