Add Metamodel::Perlable{Attribute,Class,ClassHOW}
Add roles and a class allowing runtime creation of classes that know
how to serialize their own definitions back to Perl 6 source code.

For now only handles the class/is/does and attribute declarations, but
is smart enough to serialize either a class type object or a defined
object instantiated in that class.  A defined object's serialized form
gets default assignments for each of its attributes that match the
actual current values in the object.

The roles attempt to do the right thing when handed unusual attributes,
correctly handling even BOOTSTRAPATTRs, and providing a hook for
serializing non-standard attribute traits.